+++ packages/styles/src/packages/title/index.ts
@@ -1,8 +1,12 @@
 import { style } from "./title.style.css";
-import { sprinkle, properties } from "./title.sprinkle.css";
+import {
+  sprinkle,
+  colorsProperties,
+  textAlignProperties,
+} from "./title.sprinkle.css";
 
 export const title = {
   style,
   sprinkle,
-  properties,
+  properties: { colors: colorsProperties, textAlign: textAlignProperties },
 };