+++ packages/styles/src/themes/base.css.ts
@@ -1,5 +1,98 @@
-import { createGlobalTheme } from "@vanilla-extract/css";
+import { createGlobalTheme, createThemeContract } from "@vanilla-extract/css";
 import tokens from "@nimbus-ds/tokens/dist/js/tokens";
+
+export const varsThemeBase = createThemeContract({
+  colors: {
+    primary: {
+      background: "primary.background",
+      surface: "primary.surface",
+      surfaceHighlight: "primary.surfaceHighlight",
+      interactive: "primary.interactive",
+      interactiveHover: "primary.interactiveHover",
+      interactivePressed: "primary.interactivePressed",
+      textLow: "primary.textLow",
+      textHigh: "primary.textHigh",
+    },
+    success: {
+      background: "success.background",
+      surface: "success.surface",
+      surfaceHighlight: "success.surfaceHighlight",
+      interactive: "success.interactive",
+      interactiveHover: "success.interactiveHover",
+      interactivePressed: "success.interactivePressed",
+      textLow: "success.textLow",
+      textHigh: "success.textHigh",
+    },
+    warning: {
+      background: "warning.background",
+      surface: "warning.surface",
+      surfaceHighlight: "warning.surfaceHighlight",
+      interactive: "warning.interactive",
+      interactiveHover: "warning.interactiveHover",
+      interactivePressed: "warning.interactivePressed",
+      textLow: "warning.textLow",
+      textHigh: "warning.textHigh",
+    },
+    danger: {
+      background: "danger.background",
+      surface: "danger.surface",
+      surfaceHighlight: "danger.surfaceHighlight",
+      interactive: "danger.interactive",
+      interactiveHover: "danger.interactiveHover",
+      interactivePressed: "danger.interactivePressed",
+      textLow: "danger.textLow",
+      textHigh: "danger.textHigh",
+    },
+    neutral: {
+      background: "neutral.background",
+      surface: "neutral.surface",
+      surfaceHighlight: "neutral.surfaceHighlight",
+      interactive: "neutral.interactive",
+      interactiveHover: "neutral.interactiveHover",
+      interactivePressed: "neutral.interactivePressed",
+      textLow: "neutral.textLow",
+      textHigh: "neutral.textHigh",
+    },
+  },
+  fontFamily: {
+    centranube: "fontFamily.centranube",
+  },
+  fontSize: {
+    body: {
+      small: "fontSize.body.small",
+      caption: "fontSize.body.caption",
+      base: "fontSize.body.base",
+    },
+    title: {
+      h1: "fontSize.title.h1",
+      h2: "fontSize.title.h2",
+      h3: "fontSize.title.h3",
+      h4: "fontSize.title.h4",
+      h5: "fontSize.title.h5",
+      h6: "fontSize.title.h6",
+    },
+  },
+  fontWeight: {
+    regular: "fontWeight.regular",
+    medium: "fontWeight.medium",
+    bold: "fontWeight.bold",
+  },
+  lineWeight: {
+    body: {
+      small: "lineWeight.body.small",
+      caption: "lineWeight.caption.small",
+      base: "lineWeight.base.small",
+    },
+    title: {
+      h1: "lineWeight.title.h1",
+      h2: "lineWeight.title.h2",
+      h3: "lineWeight.title.h3",
+      h4: "lineWeight.title.h4",
+      h5: "lineWeight.title.h5",
+      h6: "lineWeight.title.h6",
+    },
+  },
+});
 
 const colors = tokens.color.sys.light;
 const fontFamily = tokens.font.family;
@@ -7,7 +100,7 @@ const fontSize = tokens.font.size.sys;
 const fontWeight = tokens.font.weight;
 const lineWeight = tokens.line.height.sys;
 
-export const varsThemeBase = createGlobalTheme(":root", {
+createGlobalTheme(":root", varsThemeBase, {
   colors: {
     primary: {
       background: colors.primary.background.value as string,