+++ packages/styles/src/utils/index.ts
@@ -1,7 +1,5 @@
-import * as variantes from "./variantes";
 import * as vars from "./vars.css";
 
 export const utils = {
-  variantes,
   vars,
 };