+++ packages/styles/src/utils/variantes.ts
@@ -1,6 +1,0 @@
-export const textAlign = {
-  left: { textAlign: "left" },
-  right: { textAlign: "right" },
-  center: { textAlign: "center" },
-  justify: { textAlign: "justify" },
-} as const;