+++ packages/styles/webpack.config.js
@@ -24,7 +24,7 @@ module.exports = {
     path: path.resolve(__dirname, "dist"),
     filename: "[name].js",
     libraryTarget: "umd",
-    library: "nimbus-ds",
+    library: "@nimbus-ds/styles",
   },
   module: {
     rules: [